Refactor: changin response from DTO to ResponseEntity

+++ src/main/java/dev/raphael/cadastroApi/User/UserController.java
@@ -1,6 +1,8 @@
 package dev.raphael.cadastroApi.User;
 
 import org.apache.catalina.User;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -21,27 +23,44 @@ public class UserController {
     }
 
     @PostMapping("/")
-    public UserDTO addUser(@RequestBody UserDTO user){
-        return userService.addUser(user);
+    public ResponseEntity<String> addUser(@RequestBody UserDTO user){
+        UserDTO newUser = userService.addUser(user);
+        return ResponseEntity.status(HttpStatus.CREATED)
+                .body("User successfully created: " + newUser.getName() + " (ID): " + newUser.getId());
+
     }
 
     @GetMapping("/")
-    public List<UserDTO> getUser(){
-        return userService.listUsers();
+    public ResponseEntity<List<UserDTO>> getUser(){
+        List<UserDTO> userList = userService.listUsers();
+        return ResponseEntity.status(HttpStatus.OK)
+                .body(userList);
     }
 
     @GetMapping("/{id}")
-    public UserDTO getUserById(@PathVariable Long id){
-        return userService.findById(id);
+    public ResponseEntity<?> getUserById(@PathVariable Long id){
+        if(userService.findById(id) != null) {
+            return ResponseEntity.ok(userService.findById(id));
+        }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
     }
 
     @PatchMapping("/{id}")
-    public UserDTO editUser(@PathVariable Long id, @RequestBody UserDTO user){
-        return userService.updateUser(user, id);
+    public ResponseEntity<?> editUser(@PathVariable Long id, @RequestBody UserDTO user){
+        if(userService.findById(id) != null) {
+            return ResponseEntity.ok(userService.updateUser(user, id));
+        }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
     }
 
     @DeleteMapping("/{id}")
-    public void deleteUser(@PathVariable Long id){
-        userService.deleteById(id);
+    public ResponseEntity<String> deleteUser(@PathVariable Long id){
+        if(userService.findById(id) != null) {
+            userService.deleteById(id);
+            return ResponseEntity.ok("User successfully deleted: " + id);
+        }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                .body("User not found")
+                ;
     }
 }
